+++ .github/actions/setup-build-env/action.yml
@@ -1,0 +1,68 @@
+name: Setup Hybrid Mount Build Env
+description: Prepare the shared Android/Rust/WebUI environment for package builds
+
+inputs:
+  ndk-version:
+    description: Android NDK version to install
+    required: true
+  rust-targets:
+    description: Comma-separated Rust targets to install
+    required: true
+  node-version:
+    description: Node.js version
+    required: false
+    default: "20"
+  pnpm-version:
+    description: pnpm version
+    required: false
+    default: "10"
+
+runs:
+  using: composite
+  steps:
+    - name: Download HymoFS LKM artifacts
+      uses: actions/download-artifact@v4
+      with:
+        path: hymofs_lkm_artifacts
+        pattern: "*-hymofs-lkm"
+        merge-multiple: true
+
+    - name: Setup Android NDK
+      uses: nttld/setup-ndk@v1
+      with:
+        ndk-version: ${{ inputs.ndk-version }}
+
+    - name: Setup Rust Toolchain
+      uses: dtolnay/rust-toolchain@nightly
+      with:
+        targets: ${{ inputs.rust-targets }}
+        components: rust-src, clippy
+
+    - name: Cache Rust dependencies
+      uses: Swatinem/rust-cache@v2
+      with:
+        cache-all-crates: true
+
+    - name: Setup pnpm
+      uses: pnpm/action-setup@v5
+      with:
+        version: ${{ inputs.pnpm-version }}
+
+    - name: Setup Node.js
+      uses: actions/setup-node@v6
+      with:
+        node-version: ${{ inputs.node-version }}
+        cache: pnpm
+        cache-dependency-path: webui/pnpm-lock.yaml
+
+    - name: Install build dependencies
+      shell: bash
+      run: |
+        sudo apt update
+        sudo apt install -y gcc-multilib
+        cargo install cargo-ndk --locked
+
+    - name: Prepare APatch KPM Sources
+      shell: bash
+      run: |
+        git clone --depth 1 https://github.com/AndroidPatch/kpm.git /tmp/apatch-kpm
