+++ .github/workflows/build.yml
@@ -13,6 +13,10 @@ on:
       - 'xtask/**'
       - 'kpm/**'
       - 'build.rs'
+      - 'scripts/build-local.sh'
+      - '.github/actions/setup-build-env/**'
+      - '.github/workflows/build-hymofs-lkm.yml'
+      - '.github/workflows/hymofs-lkm.yml'
       - '.github/workflows/build.yml'
   pull_request:
     branches:
@@ -26,6 +30,10 @@ on:
       - 'xtask/**'
       - 'kpm/**'
       - 'build.rs'
+      - 'scripts/build-local.sh'
+      - '.github/actions/setup-build-env/**'
+      - '.github/workflows/build-hymofs-lkm.yml'
+      - '.github/workflows/hymofs-lkm.yml'
       - '.github/workflows/build.yml'
   workflow_dispatch:
 
@@ -50,50 +58,11 @@ jobs:
           submodules: recursive
           fetch-depth: 0
 
-      - name: Download HymoFS LKM artifacts
-        uses: actions/download-artifact@v4
-        with:
-          path: hymofs_lkm_artifacts
-          pattern: '*-hymofs-lkm'
-          merge-multiple: true
-
-      - name: Setup Android NDK
-        uses: nttld/setup-ndk@v1
+      - name: Setup Shared Build Environment
+        uses: ./.github/actions/setup-build-env
         with:
           ndk-version: r29
-
-      - name: Setup Rust Toolchain
-        uses: dtolnay/rust-toolchain@nightly
-        with:
-          targets: aarch64-linux-android
-          components: rust-src, clippy
-
-      - name: Cache Rust dependencies
-        uses: Swatinem/rust-cache@v2
-        with:
-          cache-all-crates: true
-
-      - name: Setup pnpm
-        uses: pnpm/action-setup@v5
-        with:
-          version: 10
-
-      - name: Setup Node.js
-        uses: actions/setup-node@v6
-        with:
-          node-version: '20'
-          cache: 'pnpm'
-          cache-dependency-path: webui/pnpm-lock.yaml
-
-      - name: Install build dependencies
-        run: |
-          sudo apt update
-          sudo apt install -y gcc-multilib
-          cargo install cargo-ndk --locked
-
-      - name: Prepare APatch KPM Sources
-        run: |
-          git clone --depth 1 https://github.com/AndroidPatch/kpm.git /tmp/apatch-kpm
+          rust-targets: aarch64-linux-android
 
       - name: Build All
         env:
