+++ src/mount/hymofs/mod.rs
@@ -1,0 +1,13 @@
+// Copyright 2026 Hybrid Mount Developers
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+mod common;
+mod compile;
+mod runtime;
+mod status;
+
+#[cfg(test)]
+mod tests;
+
+pub use runtime::{apply, clear_runtime_best_effort, reset_runtime, sync_runtime_config};
+pub use status::{can_operate, collect_runtime_info, hook_lines, require_live};
