+++ src/mount/mod.rs
@@ -1,6 +1,7 @@
 // Copyright 2026 Hybrid Mount Developers
 // SPDX-License-Identifier: GPL-3.0-or-later
 
+pub mod hymofs;
 pub mod magic_mount;
 pub mod node;
 pub mod overlayfs;
